refactor(algorithms): 提取社区检测和力导向布局算法至核心模块，优化画廊及朋友网络界面代码
